Team draft picks breakdown is now a component. Used in the MyTeam page and the contracts overview pages.
