+++ app/graphql/queries/league/leagueGraphQL.js
@@ -93,6 +93,15 @@ export const LEAGUE_FILTER_TEAMS_BY_OWNER_ID = gql`
                         year
                     }
                     restructureStatus
+                },
+                teamAssets {
+                    draftPicks {
+                        year
+                        picks {
+                            round
+                            value
+                        }
+                    }
                 }
             }
         }
